Refactor finances dashboard
